Revert "pruebas unitarias a los metodos Crud de JPARepository"
This reverts commit 90c0c0ae8b14ed88dfd9fec0526011a4fd8462a7.

+++ src/main/java/com/example/demo/repository/PartidaRepository.java
@@ -1,11 +1,9 @@
 package com.example.demo.repository;
 
-import com.example.demo.modelo.Mensaje;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.modelo.Partida;
 
-import java.util.List;
-
-public interface PartidaRepository extends JpaRepository<Partida,Long>{
+public interface PartidaRepository extends JpaRepository<Partida,Integer>{
+    
 }
